refactor fs compiler
